@@ -6,7 +6,11 @@ use Illuminate\Database\Eloquent\Model;
 
 class Total extends Model
 {
-    protected $dates = ['created_at', 'updated_at', 'date'];
+    protected $casts = [
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime',
+        'date' => 'date'
+    ];
 
     protected $fillable = ['date'];
 }
